Backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pyodbc==5.1.0
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0